requirements.txt
asgiref
black
dj-rest-auth
Django
django-extensions
django-filter
django-model-utils
djangorestframework
Faker
pytz
sqlparse